refactor: move CFA frontend into app package
