+++ apps/current-feature-analyzer/frontend/components/CompressionStatsCard.vue
@@ -2,7 +2,6 @@
   <el-card v-if="visible" shadow="never">
     <template #header><span class="card-title">分析统计</span></template>
     <div class="cfa-stats-grid">
-      <!-- 原始数据 -->
       <div class="cfa-stat-card" style="background:#eef6ff">
         <div class="cfa-stat-value">{{ rawPointCount }}</div>
         <div class="cfa-stat-label">原始点数</div>
@@ -15,8 +14,6 @@
         <div class="cfa-stat-value">{{ maxCurrent }}</div>
         <div class="cfa-stat-label">最大电流</div>
       </div>
-
-      <!-- 分段 -->
       <div class="cfa-stat-card" style="background:#f3e8ff">
         <div class="cfa-stat-value">{{ segmentCount }}</div>
         <div class="cfa-stat-label">压缩段数</div>
@@ -29,8 +26,6 @@
         <div class="cfa-stat-value">{{ trendCount }}</div>
         <div class="cfa-stat-label">趋势段数量</div>
       </div>
-
-      <!-- 压缩/向量化 -->
       <div class="cfa-stat-card" style="background:#eafaf1">
         <div class="cfa-stat-value">{{ compressionRatio }}</div>
         <div class="cfa-stat-label">压缩比</div>
@@ -43,8 +38,6 @@
         <div class="cfa-stat-value">{{ vectorizationRatio }}</div>
         <div class="cfa-stat-label">向量化比</div>
       </div>
-
-      <!-- 事件/冲突 -->
       <div class="cfa-stat-card" style="background:#fef3c7">
         <div class="cfa-stat-value">{{ eventCount }}</div>
         <div class="cfa-stat-label">尖峰/事件数</div>
@@ -63,7 +56,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import type { DuplicateDiagnosis, SegmentItem } from '@/api/current-feature-analyzer'
+import type { DuplicateDiagnosis, SegmentItem } from '../api/current-feature-analyzer'
 
 const props = defineProps<{
   rawPointCount: number
@@ -74,7 +67,6 @@ const props = defineProps<{
 }>()
 
 const visible = computed(() => props.rawPointCount > 0 || (props.segments?.length ?? 0) > 0)
-
 const segmentCount = computed(() => props.segments?.length ?? 0)
 const compressionRatio = computed(() => {
   if (!segmentCount.value) return '-'
@@ -88,7 +80,6 @@ const vectorizationRatio = computed(() => {
   if (!simplifiedCount.value) return '-'
   return (props.rawPointCount / simplifiedCount.value).toFixed(1)
 })
-
 const baselineMean = computed(() => {
   const v = props.globals?.baseline_mean
   return v != null ? `${Number(v).toFixed(3)} A` : '-'
@@ -97,20 +88,11 @@ const maxCurrent = computed(() => {
   const v = props.globals?.max_current
   return v != null ? `${Number(v).toFixed(2)} A` : '-'
 })
-
 const plateauKindSet = new Set(['stable', 'normal', 'off', 'plateau-low', 'plateau-mid', 'plateau-high'])
 const trendKindSet = new Set(['transition', 'rising', 'rising-fast', 'falling', 'falling-fast', 'spike', 'surge', 'drop', 'burst'])
-
-const plateauCount = computed(() => {
-  if (!props.segments) return 0
-  return props.segments.filter(s => plateauKindSet.has(s.kind)).length
-})
-const trendCount = computed(() => {
-  if (!props.segments) return 0
-  return props.segments.filter(s => trendKindSet.has(s.kind)).length
-})
+const plateauCount = computed(() => props.segments ? props.segments.filter(s => plateauKindSet.has(s.kind ?? '')).length : 0)
+const trendCount = computed(() => props.segments ? props.segments.filter(s => trendKindSet.has(s.kind ?? '')).length : 0)
 const eventCount = computed(() => props.events?.length ?? 0)
-
 const duplicateGroups = computed(() => props.duplicateDiagnosis?.duplicate_groups ?? '-')
 const duplicateRatio = computed(() => {
   if (!props.rawPointCount || !props.duplicateDiagnosis?.duplicate_groups) return '-'
