+++ apps/current-feature-analyzer/frontend/components/FileSummaryCard.vue
@@ -19,7 +19,7 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
-import type { SessionFileItem } from '@/api/current-feature-analyzer'
+import type { SessionFileItem } from '../api/current-feature-analyzer'
 
 const props = defineProps<{ file: SessionFileItem }>()
