+++ apps/current-feature-analyzer/frontend/components/StageSummaryCard.vue
@@ -20,8 +20,9 @@
 
 <script setup lang="ts">
 import { computed } from 'vue'
+import type { StageMetric } from '../api/current-feature-analyzer'
 
-const props = defineProps<{ metrics: import('@/api/current-feature-analyzer').StageMetric[] }>()
+const props = defineProps<{ metrics: StageMetric[] }>()
 
 const avgDuration = computed(() => {
   if (!props.metrics.length) return '-'
