+++ frontend/vite.config.ts
@@ -21,8 +21,14 @@ export default defineConfig(({ command }) => ({
   ],
   resolve: {
     alias: {
-      '@': fileURLToPath(new URL('./src', import.meta.url))
+      '@': fileURLToPath(new URL('./src', import.meta.url)),
+      '@apps': fileURLToPath(new URL('../apps', import.meta.url)),
+      'vue': fileURLToPath(new URL('./node_modules/vue', import.meta.url)),
+      'pinia': fileURLToPath(new URL('./node_modules/pinia', import.meta.url)),
+      'element-plus': fileURLToPath(new URL('./node_modules/element-plus', import.meta.url)),
+      '@element-plus/icons-vue': fileURLToPath(new URL('./node_modules/@element-plus/icons-vue', import.meta.url))
     },
+    dedupe: ['vue', 'pinia', 'element-plus', '@element-plus/icons-vue'],
   },
   server: {
     host: true,  // 监听所有网络接口（包括 localhost 和 127.0.0.1）
